Moving SQL common code shared in editor and dataset
