+++ lib/assets/javascripts/cartodb3/dataset/dataset-options/dataset-options-view.js
@@ -14,6 +14,7 @@ var PreviewMapView = require('./preview-map-view');
 var ActionViewEdition = require('./dataset-actions-edition-view');
 var DataSQLModel = require('../data-sql-model');
 var SQLUtils = require('../../helpers/sql-utils');
+var sqlEditorViewHelper = require('../../helpers/sql-editor-view-helper');
 var Notifier = require('../../components/notifier/notifier');
 var cdb = require('cartodb.js');
 var SQLNotifications = require('../../sql-notifications');
@@ -73,6 +74,9 @@ module.exports = CoreView.extend({
     });
 
     this._sqlModel = this._layerDefinitionModel.sqlModel;
+
+    _.extend(this, sqlEditorViewHelper);
+    this._parseSQLFunction = this._parseSQL.bind(this, this._clearSQL),
 
     SQLNotifications.track(this);
 
@@ -167,7 +171,7 @@ module.exports = CoreView.extend({
         return new DatasetEditorView({
           editorModel: self._editorModel,
           codemirrorModel: self._codemirrorModel,
-          onApplyEvent: self._parseSQL.bind(self),
+          onApplyEvent: self._parseSQLFunction,
           layerDefinitionModel: self._layerDefinitionModel,
           querySchemaModel: self._querySchemaModel
         });
@@ -178,7 +182,7 @@ module.exports = CoreView.extend({
           trackModel: self._sqlModel,
           editorModel: self._editorModel,
           queryGeometryModel: self._queryGeometryModel,
-          onApply: self._parseSQL.bind(self),
+          onApply: self._parseSQLFunction,
           previewAction: self._previewMap.bind(self),
           onClear: self._clearSQL.bind(self)
         };
@@ -192,94 +196,18 @@ module.exports = CoreView.extend({
     this._collectionPane = new TabPaneCollection(tabPaneTabs);
   },
 
-  _parseSQL: function () {
-    var appliedQuery = this._querySchemaModel.get('query');
-    var currentQuery = this._codemirrorModel.get('content');
-
-    // Remove last character if it has ';'
-    if (currentQuery && currentQuery.slice(-1) === ';') {
-      currentQuery = currentQuery.slice(0, currentQuery.length - 1);
-      this._codemirrorModel.set('content', currentQuery);
-    }
-
-    var isSameQuery = SQLUtils.isSameQuery(currentQuery, appliedQuery);
-    var altersData = SQLUtils.altersData(currentQuery);
-
-    if (currentQuery === '' || isSameQuery) {
-      return false;
-    }
-
-    if (altersData) {
-      SQLNotifications.showNotification({
-        status: 'loading',
-        info: _t('notifications.sql.alter-loading'),
-        closable: false
-      });
-
-      this._sqlModel.set('content', currentQuery);
-
-      this._SQL.execute(currentQuery, null, {
-        success: function () {
-          SQLNotifications.showNotification({
-            status: 'success',
-            info: _t('notifications.sql.alter-success'),
-            closable: true,
-            delay: Notifier.DEFAULT_DELAY
-          });
-          // Clean SQL editor (with original query) and make a fetch
-          this._clearSQL();
-        }.bind(this),
-        error: function (errors) {
-          errors = errors.responseJSON.error;
-          var parsedErrors = this._parseErrors(errors);
-          this._codemirrorModel.set('errors', parsedErrors);
-          SQLNotifications.showErrorNotification(parsedErrors);
-        }.bind(this)
-      });
-    } else {
-      this._runQuery(currentQuery, this._saveSQL.bind(this));
-    }
-  },
-
-  _runQuery: function (query, callback) {
-    this._querySchemaModel.set({
-      query: query,
-      status: 'unfetched'
-    });
-
-    this._queryGeometryModel.set({
-      query: query,
-      simple_geom: '',
-      status: 'unfetched'
-    }, { silent: true });
-
-    this._querySchemaModel.fetch({
-      success: callback
-    });
-
-    this._queryGeometryModel.fetch();
-  },
-
-  _showErrors: function (model) {
-    var errors = this._querySchemaModel.get('query_errors');
-    this._codemirrorModel.set('errors', this._parseErrors(errors));
-    this._checkClearButton();
-  },
-
-  _parseErrors: function (errors) {
-    return errors.map(function (error) {
-      return {
-        message: error
-      };
-    });
-  },
-
   _saveSQL: function () {
     var content = this._codemirrorModel.get('content');
     this._sqlModel.set('content', content);
     this._querySchemaModel.set('query_errors', []);
     this._layerDefinitionModel.save({
       sql: content
+    });
+
+    SQLNotifications.showNotification({
+      status: 'success',
+      info: _t('notifications.sql.success'),
+      closable: true
     });
 
     this._checkClearButton();
@@ -290,25 +218,12 @@ module.exports = CoreView.extend({
     });
   },
 
-  _clearSQL: function () {
-    var sql = this._defaultSQL();
-    this._codemirrorModel.set({content: sql});
-    this._querySchemaModel.set('query_errors', []);
-    this._runQuery(sql, this._saveSQL.bind(this));
-  },
-
   _defaultSQL: function (tableName) {
     return SQLUtils.getDefaultSQL(
       tableName || this._tableModel.getUnqualifiedName(),
       this._tableModel.get('permission').owner.username,
       this._userModel.isInsideOrg()
     );
-  },
-
-  _checkClearButton: function () {
-    var custom_sql = this._codemirrorModel.get('content').toLowerCase();
-    var default_sql = this._defaultSQL().toLowerCase();
-    this._clearSQLModel.set({visible: custom_sql !== default_sql});
   },
 
   _previewMap: function () {
