+++ lib/assets/javascripts/cartodb3/editor/layers/layer-content-views/data/data-view.js
@@ -16,6 +16,7 @@ var TableStats = require('../../../../components/modals/add-widgets/tablestats')
 var DataColumnsModel = require('./data-columns-model');
 var Notifier = require('../../../../components/notifier/notifier');
 var SQLUtils = require('../../../../helpers/sql-utils');
+var sqlEditorViewHelper = require('../../../../helpers/sql-editor-view-helper');
 var cdb = require('cartodb.js');
 var SQLNotifications = require('../../../../sql-notifications');
 var MetricsTracker = require('../../../../components/metrics/metrics-tracker');
@@ -78,6 +79,9 @@ module.exports = CoreView.extend({
       readonly: false
     });
 
+    _.extend(this, sqlEditorViewHelper);
+    this._parseSQLFunction = this._parseSQL.bind(this, this._applyDefaultSQL);
+
     SQLNotifications.track(this);
 
     this._configPanes();
@@ -139,96 +143,6 @@ module.exports = CoreView.extend({
     this.add_related_model(this._sqlModel);
   },
 
-  _parseSQL: function () {
-    var appliedQuery = this._querySchemaModel.get('query');
-    var currentQuery = this._codemirrorModel.get('content');
-
-    // Remove last character if it has ';'
-    if (currentQuery && currentQuery.slice(-1) === ';') {
-      currentQuery = currentQuery.slice(0, currentQuery.length - 1);
-      this._codemirrorModel.set('content', currentQuery);
-    }
-
-    var isSameQuery = SQLUtils.isSameQuery(currentQuery, appliedQuery);
-    var altersData = SQLUtils.altersData(currentQuery);
-
-    if (currentQuery === '' || isSameQuery) {
-      SQLNotifications.removeNotification();
-      return false;
-    }
-
-    if (altersData) {
-      SQLNotifications.showNotification({
-        status: 'loading',
-        info: _t('notifications.sql.alter-loading'),
-        closable: false
-      });
-
-      this._sqlModel.set('content', currentQuery);
-
-      this._SQL.execute(currentQuery, null, {
-        success: function () {
-          SQLNotifications.showNotification({
-            status: 'success',
-            info: _t('notifications.sql.alter-success'),
-            closable: true,
-            delay: Notifier.DEFAULT_DELAY
-          });
-
-          this._applyDefaultSQL();
-        }.bind(this),
-        error: function (errors) {
-          errors = errors.responseJSON.error;
-          var parsedErrors = this._parseErrors(errors);
-          this._codemirrorModel.set('errors', parsedErrors);
-          SQLNotifications.showErrorNotification(parsedErrors);
-          this._checkClearButton();
-        }.bind(this)
-      });
-    } else {
-      // Parser errors SQL here and saving
-      this._querySchemaModel.set({
-        status: 'unfetched',
-        query: currentQuery
-      });
-
-      this._queryGeometryModel.set({
-        query: currentQuery
-      });
-
-      SQLNotifications.showNotification({
-        status: 'loading',
-        info: _t('notifications.sql.applying'),
-        closable: false
-      });
-
-      var saveSQL = _.after(2, this._saveSQL.bind(this));
-      this._querySchemaModel.fetch({ success: saveSQL });
-      this._queryGeometryModel.fetch({ complete: saveSQL });
-    }
-  },
-
-  _showErrors: function (model) {
-    var errors = this._querySchemaModel.get('query_errors');
-    var hasErrors = errors && errors.length > 0;
-    this._codemirrorModel.set('errors', this._parseErrors(errors));
-    this._editorModel.set('disabled', hasErrors);
-
-    if (hasErrors) {
-      SQLNotifications.showErrorNotification(this._parseErrors(errors));
-    }
-
-    this._checkClearButton();
-  },
-
-  _parseErrors: function (errors) {
-    return errors.map(function (error) {
-      return {
-        message: error
-      };
-    });
-  },
-
   _saveSQL: function () {
     var query = this._codemirrorModel.get('content');
 
@@ -267,27 +181,12 @@ module.exports = CoreView.extend({
   _applyDefaultSQL: function () {
     var originalQuery = this._defaultSQL();
     this._codemirrorModel.set({ content: originalQuery });
+    this._clearSQLModel.set({ visible: false });
+    this._querySchemaModel.set('query_errors', []);
+
     this._sqlModel.set('content', originalQuery);
     this._userActions.saveAnalysisSourceQuery(originalQuery, this._nodeModel, this._layerDefinitionModel);
-    this._querySchemaModel.set('query_errors', []);
-    this._clearSQLModel.set({ visible: false });
     this._querySchemaModel.resetDueToAlteredData();
-  },
-
-  _clearSQL: function () {
-    var sql = this._defaultSQL();
-    this._codemirrorModel.set({
-      content: sql,
-      errors: []
-    });
-    this._clearSQLModel.set({ visible: false });
-    this._parseSQL();
-  },
-
-  _checkClearButton: function () {
-    var custom_sql = this._codemirrorModel.get('content');
-    var isDefaultQuery = SQLUtils.isSameQuery(custom_sql, this._defaultSQL());
-    this._clearSQLModel.set({ visible: !isDefaultQuery });
   },
 
   _defaultSQL: function () {
@@ -341,7 +240,7 @@ module.exports = CoreView.extend({
           layerDefinitionModel: self._layerDefinitionModel,
           querySchemaModel: self._querySchemaModel,
           codemirrorModel: self._codemirrorModel,
-          onApplyEvent: self._parseSQL.bind(self)
+          onApplyEvent: self._parseSQLFunction
         });
       },
       createActionView: function () {
@@ -354,7 +253,7 @@ module.exports = CoreView.extend({
             clearModel: self._clearSQLModel,
             applyButton: true,
             clearButton: true,
-            onApplyClick: self._parseSQL.bind(self),
+            onApplyClick: self._parseSQLFunction,
             onClearClick: self._clearSQL.bind(self)
           });
         }
